requirements.txt
# PDF Processing - LlamaIndex
llama-index-core>=0.14.0
llama-index-readers-file>=0.5.0
pypdf>=6.0.0

# LLM Extraction
openai>=2.8.0
instructor>=1.13.0

# Data Validation (included with llama-index-core but listed for clarity)
pydantic>=2.8.0

# Environment variables
python-dotenv>=1.0.0

# Date parsing
python-dateutil>=2.8.0

# Data processing and Excel export
pandas>=2.0.0
openpyxl>=3.1.0

